refactor: add theme select to user feature
