+++ src/blocks/header/HeaderUserInfo.tsx
@@ -1,5 +1,5 @@
 import * as Avatar from '@radix-ui/react-avatar'
-import { EditProfileModal } from 'features/user/components'
+import { EditProfileModal } from 'features/user/components/modals'
 import { useModal } from 'react-modal-state'
 
 import { useAppSelector } from 'hooks/redux'
