+++ src/features/kanban/board/hooks/index.ts
@@ -1,5 +1,6 @@
 export * from './useAddBoard'
 export * from './useDeleteBoard'
 export * from './useEditBoard'
+export * from './useGetAllBoards'
 export * from './useGetBoardById'
 export * from './useGetBoardId'
