+++ src/features/kanban/board/hooks/useGetAllBoards.ts
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query'
+
+import { boardService } from '../board.service'
+import { BoardCacheKeys } from '../config'
+
+export const useGetAllBoards = () =>
+  useQuery({
+    queryKey: [BoardCacheKeys.Boards],
+    queryFn: boardService.getAllBoards
+  })
