+++ src/features/user/components/index.ts
@@ -1,2 +1,1 @@
-export * from './EditProfileModal'
-export * from './NeedHelpModal'
+export * from './ThemeSelect'
