+++ src/features/user/components/modals/EditAvatar.tsx
@@ -1,7 +1,7 @@
 import type { UseMutateAsyncFunction } from '@tanstack/react-query'
 import type { AxiosError } from 'axios'
-import type { EditUserSchema } from '../user.schema'
-import type { User } from '../user.types'
+import type { EditUserSchema } from 'features/user/user.schema'
+import type { User } from 'features/user/user.types'
 
 import { useRef } from 'react'
